added big int

+++ server/models.py
@@ -228,8 +228,8 @@ class Contact(db.Model):
     
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(255), nullable=False)
-    phone = db.Column(db.Integer, nullable=False, unique=True)
-    account=db.Column(db.Integer, nullable=False, unique=True)
+    phone = db.Column(db.BigInteger, nullable=False, unique=True)
+    account=db.Column(db.BigInteger, nullable=False, unique=True)
     email = db.Column(db.String(20), nullable=True)
     user_id = db.Column(db.Integer, db.ForeignKey('users.id'))
     users = db.relationship('User', back_populates='contacts')
